File Structure Changes and Docker
