+++ frontend/src/components/Signin.jsx
@@ -1,21 +1,16 @@
 "use client";
-// import { useRouter } from 'next/router';
-// import axiosInstance from "../../../helpers/axiosInstance";
+
 import Image from "next/image";
-import myImg from "../../../public/zine.png";
-import GoogleLoginButton from '../../components/GoogleLoginButton';
-
-
-
+import myImg from "../../public/zine.png";
 import { useState } from "react";
 
-export default function SignUp({ onTabChange }) {
-    const [username, setUsername] = useState("");
+import GoogleLoginButton from './GoogleLoginButton';
+
+export default function SignIn() {
+
     const [email, setEmail] = useState("");
     const [password, setPassword] = useState("");
-    const [confirmPassword, setConfirmPassword] = useState("")
     const [errorMessage, setErrorMessage] = useState(""); // Error message state
-    const [successMessage, setSuccessMessage] = useState(""); // Add success message state
 
     // Email validation regex
     const validateEmail = (email) => {
@@ -23,14 +18,11 @@ export default function SignUp({ onTabChange }) {
         return emailRegex.test(email);
     };
 
-    // Handle form submission from Login Button
-    const handleRegister = async (e) => {
+    // Handle form submission from Sign up Button
+    const handleLogin = async (e) => {
         e.preventDefault(); // Prevent default form submission behavior
 
-        setErrorMessage("");
-        setSuccessMessage(""); // Clear any existing success message
-
-        if (!username || !email || !password) {
+        if (!email || !password) {
             setErrorMessage("Please enter both email/username and password.");
             return;
         }
@@ -41,18 +33,13 @@ export default function SignUp({ onTabChange }) {
             return;
         }
 
-        if (!confirmPassword || password != confirmPassword) {
-            setErrorMessage("Password does not match!");
-            return;
-        }
         const payload = {
-            name: username,
             email: email,
             password: password,
         };
 
         try {
-            const response = await fetch("http://localhost:8080/consumer/register", {
+            const response = await fetch("http://localhost:8080/consumer/login", {
                 method: "POST",
                 headers: {
                     "Content-Type": "application/json",
@@ -60,36 +47,27 @@ export default function SignUp({ onTabChange }) {
                 body: JSON.stringify(payload),
             });
 
-
             if (!response.ok) {
                 const errorData = await response.json();
-                setErrorMessage(errorData.message || "Sign up failed.");
+                setErrorMessage(errorData.message || "Login failed.");
                 return;
             }
 
             const data = await response.json();
-
-            // Show success message and switch to signin tab
-            setSuccessMessage("Registration successful! Please sign in with your credentials.");
-
-            // After showing success message, switch to signin tab
-            setTimeout(() => {
-                if (onTabChange) {
-                    onTabChange("signin"); // Switch to signin tab
-                }
-            }, 2000); // 2 second delay
-
+            // Store tokens and redirect or update UI as needed
+            document.cookie = "accessToken=" + data.access_token + "; path=/; secure; samesite=strict";
+            document.cookie = "refreshToken=" + data.refresh_token + "; path=/; secure; samesite=strict";
+            document.cookie = "expires_at=" + data.expires_at + "; path=/; secure; samesite=strict";
+            window.location.href = "/home"; // Redirect upon success
         } catch (error) {
-            console.error("Error during register:", error);
+            console.error("Error during login:", error);
             setErrorMessage("An error occurred. Please try again.");
         }
     };
-
-
     return (
         <div className="bg-gray-50 flex flex-col items-center justify-center space-y-4">
             {/* Main Login Card */}
-            <div className="w-full max-w-s bg-white border border-gray-300 rounded-md p-8">
+            <div className="w-full max-w-s bg-white border border-gray-300 rounded-lg p-8">
                 {/* Zine Logo */}
                 <div className="flex justify-center mb-6">
                     <Image src={myImg} width={70} height={70} alt="Zine" />
@@ -105,25 +83,7 @@ export default function SignUp({ onTabChange }) {
                         </button>
                     </div>
                 )}
-                {successMessage && (
-                    <div data-testid="success-message" className="mb-4 flex items-center justify-between bg-green-100 border border-green-400 text-green-700 px-4 py-2 rounded text-sm">
-                        <span>{successMessage}</span>
-                        <button
-                            onClick={() => setSuccessMessage('')}
-                            className="ml-4 font-bold focus:outline-none"
-                        >
-                            X
-                        </button>
-                    </div>
-                )}
                 <form>
-                    <input
-                        type="text"
-                        placeholder="Enter Username"
-                        className="mb-3 w-full px-3 py-2 border border-gray-300 rounded text-sm text-black focus:outline-none focus:ring-1 focus:ring-gray-400"
-                        value={username}
-                        onChange={(e) => setUsername(e.target.value)}
-                    />
                     <input
                         type="text"
                         placeholder="Email"
@@ -138,19 +98,13 @@ export default function SignUp({ onTabChange }) {
                         value={password}
                         onChange={(e) => setPassword(e.target.value)}
                     />
-                    <input
-                        type="password"
-                        placeholder="Confirm Password"
-                        className="mb-3 w-full px-3 py-2 border border-gray-300 rounded text-sm text-black focus:outline-none focus:ring-1 focus:ring-gray-400"
-                        value={confirmPassword}
-                        onChange={(e) => setConfirmPassword(e.target.value)}
-                    />
+
                     <button
                         type="button"
                         className="w-full bg-blue-500 text-white py-2 rounded font-semibold text-sm hover:bg-blue-600 transition"
-                        onClick={handleRegister}
+                        onClick={handleLogin} // Call handleLogin function on click
                     >
-                        Sign up
+                        Log in
                     </button>
                     <div className="text-center text-sm text-gray-500 my-4">
                         OR
@@ -159,6 +113,11 @@ export default function SignUp({ onTabChange }) {
                         <GoogleLoginButton setErrorMessage={setErrorMessage} />
                     </div>
                 </form>
+                <div className="mt-6 text-center">
+                    <a href="/forgot" className="text-sm text-blue-600 font-semibold">
+                        Forgot password?
+                    </a>
+                </div>
             </div>
         </div>
     );
